pyproject.toml
# poetry add <package name as PyPi>  # Add new package. Then select the version from the list
# poetry install  # Install all dependencies
# poetry update  # Update all dependencies
# poetry lock  # Lock the dependencies
# poetry show  # Show all dependencies
# poetry show <package name>  # Show a specific package
# poetry remove <package name>  # Remove a package
# poetry lock --no-update  # Regenerate the lock file to make it consistent with the current state of the project
# poetry install --no-root  # Install all dependencies without installing the project itself
# 
[tool.poetry]
name = "ecpa"
version = "0.2.0"
description = "Surface Vessels Navigation Prediction"
authors = ["Tom Beyer <stu222518@mail.uni-kiel.de>", "Shang Gao <sgao@uni-kassel.de>", "Ghassan Al-Falouji <gaf@informatik.uni-kiel.de>"]
license = "MIT"
readme = "README.md"

[tool.setuptools_scm]
write_to = "./_version.py"

[tool.poetry.dependencies]
python = ">=3.12, <3.13"
numpy = ">=1.26.2"
SciPy = ">=1.11.4"
pandas = ">=2.1.4"
pandas-geojson = ">=1.2.0"
sklearn-pandas = ">=2.2.0"
pandas-datareader = ">=0.10.0"
h5py = ">=3.10.0"
pyyaml = ">=6.0.1"
seaborn = ">=0.13.1"
pyqt6 = ">=6.6.1"
ipython = ">=8.19.0"
ipykernel = ">=6.28.0"
ipywidgets = ">=8.1.1"
jupyter = ">=1.0.0"
jupyterlab = ">=4.0.10"
notebook = ">=7.0.6"
nb-conda-store-kernels = ">=0.1.5"
jupyter-contrib-nbextensions = ">=0.7.0"
jupyter-nbextensions-configurator = ">=0.6.3"
nbclean = ">=0.3.2"
nbresuse = ">=0.4.0"
autopep8 = ">=2.0.4"
pygments = ">=2.17.2"
scikit-learn = ">=1.3.2"
scikit-learn-intelex = ">=2024.0.1"
pillow = ">=10.1.0"
requests = ">=2.31.0"
tqdm = ">=4.66.1"
regex = ">=2023.12.25"
dotsi = ">=0.0.3"
pyais = ">=2.6.0"
geopandas = ">=0.14.1"
movingpandas = ">=0.17.2"
flask = ">=3.0.0"
contextily = ">=1.5.0"
pyarrow = ">=14.0.2"
folium = ">=0.15.1"
jupyter-bokeh = ">=3.0.7"
sklearn-model = ">=0.0.6"
geoplot = ">=0.5.1"
vincenty = ">=0.1.4"
tornado = ">=6.4"
stonesoup = ">=1.1"
osmnx = ">=1.8.1"
pyproj = ">=3.6.1"
geomag = ">=0.9.2015"
snowflake-connector-python = ">=3.6.0"
haversine = ">=2.8.0"
fastdtw = ">=0.3.4"
tsfel = ">=0.1.6"
geoviews = ">=1.11.0"
hdbscan = ">=0.8.33"
holoviews = ">=1.18.1"
fiona = ">=1.9.5"
shapely = ">=2.0.2"
rtree = ">=1.1.0"
geonetworkx = ">=0.5.3"
hvplot = ">=0.9.1"
torch = { version = ">=2.0.1", source="torch"}
torchaudio = { version = ">=2.0.2", source="torch"}
torchvision = { version = ">=0.15.2", source="torch"}
opencv-python = ">=4.9"
statsmodels = ">=0.14.1"
filterpy = ">=1.4.5"
xgboost = ">=2.0.3"
scienceplots = ">=2.1.1"
ipympl = ">=0.9.4"
mpl-tools = ">=0.4.0"
# basemap = ">=1.4.1"
cartopy = ">=0.23.0"
pykalman = "^0.9.7"
microprediction = "^1.2.0"
geopy = "^2.4.1"
darts = "^0.30.0"
catboost = "^1.2.5"
prophet = "^1.1.5"
lightgbm = "^4.4.0"
dask = {extras = ["dataframe"], version = "^2024.6.2"}


[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu121"  # "https://download.pytorch.org/whl/cu117"  # Adjust for CUDA version if needed
priority = "explicit"


[tool.peotry.extras]
dev = [
    "isort",
    "black",
    "flake8",
    "mypy",
    "pytest",
    "pytest-cov",
    "pre-commit",  # Consider this for automated code formatting and linting
]

[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]  # ["poetry-core"]
build-backend = "setuptools.build_meta"  # "poetry.core.masonry.api"

